requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
openai
redis
itsdangerous
python-multipart
numpy